+++ hm-dianping/src/main/java/com/hmdp/service/IShopTypeService.java
@@ -1,5 +1,6 @@
 package com.hmdp.service;
 
+import com.hmdp.dto.Result;
 import com.hmdp.entity.ShopType;
 import com.baomidou.mybatisplus.extension.service.IService;
 
@@ -12,5 +13,10 @@ import com.baomidou.mybatisplus.extension.service.IService;
  * @since 2021-12-22
  */
 public interface IShopTypeService extends IService<ShopType> {
+
+    /**
+     * 查询商铺分类列表（带 Redis 缓存）
+     */
+    Result queryTypeList();
 
 }